refactor: Complete deep rebranding - remove Void upstream traces
- Replace VoidEditor/void-editor/void.so with Orcide/orcest.ai
- Rewrite README with Orcest AI ecosystem branding
- Update extension configs and source files
- Keep void/ directory structure intact (required for VS Code contrib system)

Part of the Orcest AI ecosystem transformation.
