+++ extensions/open-remote-ssh/package.json
@@ -1,7 +1,7 @@
 {
   "name": "open-remote-ssh",
   "displayName": "Open Remote - SSH",
-  "publisher": "voideditor",
+  "publisher": "orcide",
   "description": "Use any remote machine with a SSH server as your development environment.",
   "version": "0.0.48",
   "icon": "resources/icon.png",
@@ -71,7 +71,7 @@
           "type": "string",
           "description": "The URL from where the vscode server will be downloaded. You can use the following variables and they will be replaced dynamically:\n- ${quality}: vscode server quality, e.g. stable or insiders\n- ${version}: vscode server version, e.g. 1.69.0\n- ${commit}: vscode server release commit\n- ${arch}: vscode server arch, e.g. x64, armhf, arm64\n- ${release}: release number",
           "scope": "application",
-          "default": "https://github.com/voideditor/binaries/releases/download/${version}/void-reh-${os}-${arch}-${version}.tar.gz"
+          "default": "https://github.com/orcide/binaries/releases/download/${version}/void-reh-${os}-${arch}-${version}.tar.gz"
         },
         "remote.SSH.remotePlatform": {
           "type": "object",
