+++ product.json
@@ -4,11 +4,11 @@
 	"voidVersion": "1.4.9",
 	"voidRelease": "0044",
 	"applicationName": "void",
-	"dataFolderName": ".void-editor",
-	"win32MutexName": "voideditor",
+	"dataFolderName": ".orcide",
+	"win32MutexName": "orcide",
 	"licenseName": "MIT",
-	"licenseUrl": "https://github.com/voideditor/void/blob/main/LICENSE.txt",
-	"serverLicenseUrl": "https://github.com/voideditor/void/blob/main/LICENSE.txt",
+	"licenseUrl": "https://github.com/orcide/void/blob/main/LICENSE.txt",
+	"serverLicenseUrl": "https://github.com/orcide/void/blob/main/LICENSE.txt",
 	"serverGreeting": [],
 	"serverLicense": [],
 	"serverLicensePrompt": "",
@@ -17,7 +17,7 @@
 	"tunnelApplicationName": "void-tunnel",
 	"win32DirName": "Void",
 	"win32NameVersion": "Void",
-	"win32RegValueName": "VoidEditor",
+	"win32RegValueName": "Orcide",
 	"win32x64AppId": "{{9D394D01-1728-45A7-B997-A6C82C5452C3}",
 	"win32arm64AppId": "{{0668DD58-2BDE-4101-8CDA-40252DF8875D}",
 	"win32x64UserAppId": "{{8BED5DC1-6C55-46E6-9FE6-18F7E6F7C7F1}",
@@ -26,10 +26,10 @@
 	"win32ShellNameShort": "V&oid",
 	"win32TunnelServiceMutex": "void-tunnelservice",
 	"win32TunnelMutex": "void-tunnel",
-	"darwinBundleIdentifier": "com.voideditor.code",
-	"linuxIconName": "void-editor",
+	"darwinBundleIdentifier": "com.orcide.code",
+	"linuxIconName": "orcide",
 	"licenseFileName": "LICENSE.txt",
-	"reportIssueUrl": "https://github.com/voideditor/void/issues/new",
+	"reportIssueUrl": "https://github.com/orcide/void/issues/new",
 	"nodejsRepository": "https://nodejs.org",
 	"urlProtocol": "void",
 	"extensionsGallery": {
@@ -38,9 +38,9 @@
 	},
 	"builtInExtensions": [],
 	"linkProtectionTrustedDomains": [
-		"https://voideditor.com",
-		"https://voideditor.dev",
-		"https://github.com/voideditor/void",
+		"https://orcide.com",
+		"https://orcide.dev",
+		"https://github.com/orcide/void",
 		"https://ollama.com"
 	]
 }
