+++ src/vs/workbench/contrib/void/browser/autocompleteService.ts
@@ -28,7 +28,7 @@ import { IConvertToLLMMessageService } from './convertToLLMMessageService.js';
 const allLinebreakSymbols = ['\r\n', '\n']
 const _ln = isWindows ? allLinebreakSymbols[0] : allLinebreakSymbols[1]
 
-// The extension this was called from is here - https://github.com/voideditor/void/blob/autocomplete/extensions/void/src/extension/extension.ts
+// The extension this was called from is here - https://github.com/orcide/void/blob/autocomplete/extensions/void/src/extension/extension.ts
 
 
 /*
