+++ src/vs/workbench/contrib/void/browser/react/tsup.config.js
@@ -7,7 +7,7 @@ import { defineConfig } from 'tsup'
 
 export default defineConfig({
 	entry: [
-		'./src2/void-editor-widgets-tsx/index.tsx',
+		'./src2/orcide-widgets-tsx/index.tsx',
 		'./src2/sidebar-tsx/index.tsx',
 		'./src2/void-settings-tsx/index.tsx',
 		'./src2/void-tooltip/index.tsx',
