+++ src/vs/workbench/contrib/void/browser/voidSelectionHelperWidget.ts
@@ -11,7 +11,7 @@ import { IEditorContribution } from '../../../../editor/common/editorCommon.js';
 import { Selection } from '../../../../editor/common/core/selection.js';
 import { RunOnceScheduler } from '../../../../base/common/async.js';
 import * as dom from '../../../../base/browser/dom.js';
-import { mountVoidSelectionHelper } from './react/out/void-editor-widgets-tsx/index.js';
+import { mountVoidSelectionHelper } from './react/out/orcide-widgets-tsx/index.js';
 import { IInstantiationService } from '../../../../platform/instantiation/common/instantiation.js';
 import { IVoidSettingsService } from '../common/voidSettingsService.js';
 import { EditorOption } from '../../../../editor/common/config/editorOptions.js';
