+++ src/vs/workbench/contrib/void/electron-main/voidUpdateMainService.ts
@@ -95,7 +95,7 @@ export class VoidMainUpdateService extends Disposable implements IVoidUpdateServ
 
 	private async _manualCheckGHTagIfDisabled(explicit: boolean): Promise<VoidCheckUpdateRespose> {
 		try {
-			const response = await fetch('https://api.github.com/repos/voideditor/binaries/releases/latest');
+			const response = await fetch('https://api.github.com/repos/orcide/binaries/releases/latest');
 
 			const data = await response.json();
 			const version = data.tag_name;
